Move useFetch hook into utils folder
