+++ src/components/Menu.jsx
@@ -1,4 +1,4 @@
-import { useFetch } from "../useFetch";
+import { useFetch } from "../utils/useFetch";
 import "../table.css";
 import { CategoryTable } from "./CategoryTable";
